Add RE (Regular Expression) extention (foreign function 're-match')

+++ ext/re/regcomp.c
@@ -4,7 +4,7 @@
 #include <ctype.h>
 #include <limits.h>
 #include <stdlib.h>
-#include <regex.h>
+#include "regex.h"
 
 #include "utils.h"
 #include "regex2.h"
@@ -1062,7 +1062,7 @@ freeset(p, cs)
 register struct parse *p;
 register cset *cs;
 {
-	register int i;
+	register unsigned int i;
 	register cset *top = &p->g->sets[p->g->ncsets];
 	register size_t css = (size_t)p->g->csetsize;
 
@@ -1088,7 +1088,7 @@ register struct parse *p;
 register cset *cs;
 {
 	register uch h = cs->hash;
-	register int i;
+	register size_t i;
 	register cset *top = &p->g->sets[p->g->ncsets];
 	register cset *cs2;
 	register size_t css = (size_t)p->g->csetsize;
@@ -1121,7 +1121,7 @@ firstch(p, cs)
 register struct parse *p;
 register cset *cs;
 {
-	register int i;
+	register size_t i;
 	register size_t css = (size_t)p->g->csetsize;
 
 	for (i = 0; i < css; i++)
@@ -1140,7 +1140,7 @@ nch(p, cs)
 register struct parse *p;
 register cset *cs;
 {
-	register int i;
+	register size_t i;
 	register size_t css = (size_t)p->g->csetsize;
 	register int n = 0;
